+++ docs/source/changelog.rst
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file.
 `Unreleased`_
 -------------
 
+`0.4.0`_ -- 2018-07-28
+----------------------
+
 Added
 ^^^^^
 
@@ -98,7 +101,8 @@ Changed
 * Initial release.
 
 
-.. _Unreleased: https://github.com/Bogdanp/molten/compare/v0.3.3...HEAD
+.. _Unreleased: https://github.com/Bogdanp/molten/compare/v0.4.0...HEAD
+.. _0.4.0: https://github.com/Bogdanp/molten/compare/v0.3.3...v0.4.0
 .. _0.3.3: https://github.com/Bogdanp/molten/compare/v0.3.2...v0.3.3
 .. _0.3.2: https://github.com/Bogdanp/molten/compare/v0.3.1...v0.3.2
 .. _0.3.1: https://github.com/Bogdanp/molten/compare/v0.3.0...v0.3.1